F1.0-T09+T10: puertas de CI y andamiaje de pruebas
PUERTAS (T09)
ESLint con typescript-eslint, no Biome. La razon es concreta:
no-floating-promises. En un sistema donde el cobro, la comanda y el
stock son transacciones (R10), un await olvidado no falla — sigue y deja
la transaccion a medias. Es el defecto P0-03 de las dos fuentes. Biome
todavia no tiene reglas con informacion de tipos.

Tambien: no-misused-promises, only-throw-error, no-empty sin excepcion
para catch (R12), no-console, y parseFloat prohibido para importes (R15).

Las 5 prohibiciones de capas de 04-ARQUITECTURA §2, verificadas una por
una creando el import prohibido a proposito.

DEFECTO ENCONTRADO POR LA META-PRUEBA: la quinta prohibicion (historico/)
aplicaba a **/* y, al ir despues, DESACTIVABA EN SILENCIO las otras
cuatro. En la configuracion plana de ESLint dos bloques que tocan el
mismo archivo y la misma regla no se suman: gana el ultimo. Sin crear el
fallo a proposito, la regla de dependencia entre capas habria estado
apagada desde el primer dia sin que nada lo dijera.

verificar-residuos.mjs no escribe los patrones prohibidos literalmente:
los compone. Si estuvieran escritos tal cual, el escaner se encontraria a
si mismo y habria que exceptuarlo — y una excepcion es por donde se
empieza a perder una regla. Lo mismo en el workflow y en
verificar-historico.

El workflow de GitHub Actions queda escrito y sin ejecutar (A-35, sin
remoto), con las acciones fijadas a SHA completo. Cada paso invoca el
mismo script que corre en local, y esos si estan verificados.

ANDAMIAJE (T10)
- Inyeccion de fallos por NOMBRE de paso, no por posicion: un indice se
  rompe al agregar un paso intermedio y la prueba seguiria pasando
  interrumpiendo otra cosa. Si el paso no existe, revienta.
- Postgres real para integracion, con contenedor efimero en tmpfs. Si no
  hay base, FALLA con instrucciones: no se salta. Una suite verde sin las
  pruebas de integracion da una seguridad que no existe.
- Datos sinteticos de 13-PRUEBAS §7: nombres con HTML, comillas, saltos,
  emoji y unicode largo; reloj controlado; ids deterministas.
- Playwright con 5 pruebas x 2 dispositivos. Automatiza la revision de
  zoom a 100/125/200 % que hasta ahora era manual.

DOS DEFECTOS QUE ENCONTRO LA E2E:
1. Tailwind 4 NO genera `h-[var(--x)]` — descarta el var() desnudo entre
   corchetes; su sintaxis es `h-(--x)`. La clase quedaba en el HTML, no
   existia en el CSS, y la perilla de densidad no hacia nada. Sin fallar
   nada.
2. shadcn usa `transition-all`, que anima la ALTURA. Viola la regla de
   rendimiento del proyecto y hacia que el control no respondiera al
   cambio de densidad. Sustituido por la lista explicita de propiedades.

La prueba de densidad primero comparaba ordenes ("compacta < normal") y
NO detectaba un h-10 fijo, porque la escala de espaciado de Tailwind esta
aliada a nuestros tokens y tambien encoge. Se cambio a alturas EXACTAS
(48/40/32). Con la mutacion h-10 da 50 y falla.

+++ packages/domain/src/dinero/aritmetica.ts
@@ -27,7 +27,7 @@ export function restar(minuendo: Centavos, sustraendo: Centavos): Centavos {
 }
 
 export function negar(monto: Centavos): Centavos {
-  return -monto as Centavos;
+  return -(monto as bigint) as Centavos;
 }
 
 /** -1 si a < b, 0 si son iguales, 1 si a > b. Sirve para ordenar. */
